reco-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-ratelimit>=4.1
django-q2>=1.6
mysqlclient>=2.2
boto3>=1.34
opensearch-py>=2.4
openai-whisper>=20231117
ffmpeg-python>=0.2
gunicorn>=21.2
django-storages>=1.14
Pillow>=10.2
openpyxl>=3.1
pydantic>=2.5
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
python-json-logger>=2.0
django-health-check>=3.18
python-dotenv>=1.0
